requirements.txt
Flask==3.0.3
spacy==3.7.5
sentence-transformers==3.0.1
python-dotenv==1.0.1
dateparser==1.2.0
pdfplumber==0.11.1

# --- Pinned versions for Python 3.12 compatibility ---
numpy==1.26.4
scipy==1.12.0
gensim==4.3.2
nltk==3.8.1
textblob==0.18.0.post0